+++ src/Clay.WebApi/Migrations/WebApiDbContextModelSnapshot.cs
@@ -59,7 +59,7 @@ namespace Clay.WebApi.Migrations
 
             modelBuilder.Entity("Clay.WebApi.CardGroup", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
@@ -68,7 +68,7 @@ namespace Clay.WebApi.Migrations
                     b.Property<string>("Description")
                         .IsRequired();
 
-                    b.Property<long?>("PropertyId");
+                    b.Property<int?>("PropertyId");
 
                     b.HasKey("Id");
 
@@ -81,15 +81,15 @@ namespace Clay.WebApi.Migrations
 
             modelBuilder.Entity("Clay.WebApi.CardGroupLock", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<int?>("AuditId");
 
-                    b.Property<long?>("CardGroupId");
+                    b.Property<int?>("CardGroupId");
 
-                    b.Property<long?>("LockId");
+                    b.Property<int?>("LockId");
 
                     b.HasKey("Id");
 
@@ -133,7 +133,7 @@ namespace Clay.WebApi.Migrations
 
             modelBuilder.Entity("Clay.WebApi.Lock", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
@@ -149,7 +149,7 @@ namespace Clay.WebApi.Migrations
 
                     b.Property<int>("LockState");
 
-                    b.Property<long?>("PropertyId");
+                    b.Property<int?>("PropertyId");
 
                     b.HasKey("Id");
 
@@ -164,7 +164,7 @@ namespace Clay.WebApi.Migrations
 
             modelBuilder.Entity("Clay.WebApi.LockCard", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
@@ -172,7 +172,7 @@ namespace Clay.WebApi.Migrations
 
                     b.Property<int?>("CardId");
 
-                    b.Property<long?>("LockId");
+                    b.Property<int?>("LockId");
 
                     b.HasKey("Id");
 
@@ -187,7 +187,7 @@ namespace Clay.WebApi.Migrations
 
             modelBuilder.Entity("Clay.WebApi.LockEvent", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
@@ -201,7 +201,7 @@ namespace Clay.WebApi.Migrations
 
                     b.Property<int>("EventType");
 
-                    b.Property<long>("LockId");
+                    b.Property<int>("LockId");
 
                     b.HasKey("Id");
 
@@ -218,7 +218,7 @@ namespace Clay.WebApi.Migrations
 
             modelBuilder.Entity("Clay.WebApi.Property", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
